chore: gorilla mux native test passing
