+++ gin/gin_test.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"encoding/json"
 	"errors"
-	"fmt"
 	"io"
 	"net/http"
 	"net/http/httptest"
@@ -188,65 +187,6 @@ func TestOutgoingRequestGin(t *testing.T) {
 	defer ts.Close()
 
 	_, err := req.Get(ts.URL + "/slug-value/test")
-	assert.NoError(t, err)
-	assert.True(t, publishCalled)
-}
-
-func TestErrorReporting(t *testing.T) {
-	client := &apt.Client{}
-	client.SetConfig(&apt.Config{
-		RedactHeaders:      []string{"X-Api-Key", "Accept-Encoding"},
-		RedactResponseBody: apt.ExampleDataRedaction,
-	})
-
-	var publishCalled bool
-	client.PublishMessage = func(ctx context.Context, payload apt.Payload) error {
-		// x, _ := json.MarshalIndent(payload, "", "\t")
-		// fmt.Println(string(x))
-		assert.NotEmpty(t, payload.Errors)
-		assert.Equal(t, "wrapper from err2 Example Error value", payload.Errors[0].Message)
-		assert.Equal(t, "Example Error value", payload.Errors[0].RootErrorMessage)
-		assert.Equal(t, "*fmt.wrapError", payload.Errors[0].ErrorType)
-		assert.Equal(t, "*errors.errorString", payload.Errors[0].RootErrorType)
-
-		assert.Equal(t, "POST", payload.Method)
-		assert.Equal(t, "/test", payload.URLPath)
-		publishCalled = true
-		return nil
-	}
-
-	handlerFn := func(w http.ResponseWriter, r *http.Request) {
-		err1 := fmt.Errorf("Example Error %v", "value")
-
-		w.WriteHeader(http.StatusAccepted)
-		w.Write([]byte(`{"key":"value"}`))
-
-		err2 := fmt.Errorf("wrapper from err2 %w", err1)
-		ReportError(r.Context(), err2)
-	}
-
-	ts := httptest.NewServer(client.Middleware(http.HandlerFunc(handlerFn)))
-	defer ts.Close()
-
-	outClient := &apt.Client{}
-	outClient.SetConfig(&apt.Config{})
-
-	outClient.PublishMessage = func(ctx context.Context, payload apt.Payload) error {
-		assert.Equal(t, "/test?param1=abc&param2=123", payload.RawURL)
-		assert.Equal(t, http.StatusAccepted, payload.StatusCode)
-		assert.Greater(t, payload.Duration, 1000*time.Nanosecond)
-		assert.Equal(t, apt.GoOutgoing, payload.SdkType)
-		return nil
-	}
-
-	_, err := req.Post(ts.URL+"/test",
-		req.Param{"param1": "abc", "param2": 123},
-		req.Header{
-			"Content-Type": "application/json",
-			"X-API-KEY":    "past-3",
-		},
-		req.BodyJSON(apt.ExampleData2),
-	)
 	assert.NoError(t, err)
 	assert.True(t, publishCalled)
 }
