+++ gorilla/gorilla.go
@@ -1,15 +1,17 @@
-package apitoolkitnative
+package apitoolkitgorilla
 
 import (
 	"bytes"
 	"context"
 	"io"
+	"log"
 	"net/http"
 	"net/http/httptest"
 	"time"
 
 	apt "github.com/apitoolkit/apitoolkit-go"
 	"github.com/google/uuid"
+	"github.com/gorilla/mux"
 )
 
 func NewClient(ctx context.Context, conf apt.Config) (*apt.Client, error) {
@@ -34,8 +36,8 @@ func WithRedactResponseBody(paths ...string) apt.RoundTripperOption {
 	return apt.WithRedactResponseBody(paths...)
 }
 
-// Middleware collects request, response parameters and publishes the payload
-func Middleware(c *apt.Client) func(http.Handler) http.Handler {
+// GorillaMuxMiddleware is for the gorilla mux routing library and collects request, response parameters and publishes the payload
+func GorillaMuxMiddleware(c *apt.Client) func(next http.Handler) http.Handler {
 	return func(next http.Handler) http.Handler {
 		return http.HandlerFunc(func(res http.ResponseWriter, req *http.Request) {
 			msgID := uuid.Must(uuid.NewRandom())
@@ -55,7 +57,6 @@ func Middleware(c *apt.Client) func(http.Handler) http.Handler {
 			next.ServeHTTP(rec, req)
 
 			recRes := rec.Result()
-			// io.Copy(res, recRes.Body)
 			for k, v := range recRes.Header {
 				for _, vv := range v {
 					res.Header().Add(k, vv)
@@ -65,17 +66,26 @@ func Middleware(c *apt.Client) func(http.Handler) http.Handler {
 			res.WriteHeader(recRes.StatusCode)
 			res.Write(resBody)
 
+			route := mux.CurrentRoute(req)
+			pathTmpl, _ := route.GetPathTemplate()
+			vars := mux.Vars(req)
 			config := c.GetConfig()
 
-			payload := c.BuildPayload(apt.GoDefaultSDKType, start,
+			payload := c.BuildPayload(apt.GoGorillaMux, start,
 				req, recRes.StatusCode,
-				reqBuf, resBody, recRes.Header, nil, req.URL.Path,
+				reqBuf, resBody, recRes.Header, vars, pathTmpl,
 				config.RedactHeaders, config.RedactRequestBody, config.RedactResponseBody,
 				errorList,
 				msgID,
 				nil,
 			)
-			c.PublishMessage(req.Context(), payload)
+
+			err := c.PublishMessage(req.Context(), payload)
+			if err != nil {
+				if config.Debug {
+					log.Println("APIToolkit: unable to publish request payload to pubsub.")
+				}
+			}
 		})
 	}
 }
