+++ gorilla/gorilla_test.go
@@ -1,4 +1,4 @@
-package apitoolkit
+package apitoolkitgorilla
 
 import (
 	"context"
@@ -9,94 +9,22 @@ import (
 	"testing"
 	"time"
 
+	apt "github.com/apitoolkit/apitoolkit-go"
 	"github.com/gorilla/mux"
 	"github.com/imroc/req"
 	"github.com/stretchr/testify/assert"
 )
 
-func TestNativeGoMiddleware(t *testing.T) {
-	client := &Client{
-		config: &Config{
-			RedactHeaders:      []string{"X-Api-Key", "Accept-Encoding"},
-			RedactResponseBody: ExampleDataRedaction,
-		},
-	}
-	var publishCalled bool
-	client.PublishMessage = func(ctx context.Context, payload Payload) error {
-		assert.Equal(t, "POST", payload.Method)
-		assert.Equal(t, "/test", payload.URLPath)
-		assert.Equal(t, map[string]string(nil), payload.PathParams)
-		assert.Equal(t, map[string][]string{
-			"param1": {"abc"},
-			"param2": {"123"},
-		}, payload.QueryParams)
-
-		assert.Equal(t, map[string][]string{
-			"Accept-Encoding": {"gzip"},
-			"Content-Length":  {"437"},
-			"Content-Type":    {"application/json"},
-			"User-Agent":      {"Go-http-client/1.1"},
-			"X-Api-Key":       {"past-3"},
-		}, payload.RequestHeaders)
-		assert.Equal(t, map[string][]string{
-			"Content-Type": {"application/json"},
-			"X-Api-Key":    {"applicationKey"},
-		}, payload.ResponseHeaders)
-		assert.Equal(t, "/test?param1=abc&param2=123", payload.RawURL)
-		assert.Equal(t, http.StatusAccepted, payload.StatusCode)
-		assert.Greater(t, payload.Duration, 1000*time.Nanosecond)
-		assert.Equal(t, GoDefaultSDKType, payload.SdkType)
-
-		reqData, _ := json.Marshal(ExampleData2)
-		respData, _ := json.Marshal(ExampleDataRedacted)
-
-		assert.Equal(t, reqData, payload.RequestBody)
-		assert.Equal(t, respData, payload.ResponseBody)
-
-		publishCalled = true
-		return nil
-	}
-
-	handlerFn := func(w http.ResponseWriter, r *http.Request) {
-		body, err := io.ReadAll(r.Body)
-		assert.NoError(t, err)
-		assert.NotEmpty(t, body)
-
-		jsonByte, err := json.Marshal(ExampleData)
-		assert.NoError(t, err)
-
-		w.Header().Add("Content-Type", "application/json")
-		w.Header().Add("X-API-KEY", "applicationKey")
-		w.WriteHeader(http.StatusAccepted)
-		w.Write(jsonByte)
-	}
-
-	ts := httptest.NewServer(client.Middleware(http.HandlerFunc(handlerFn)))
-	defer ts.Close()
-
-	_, err := req.Post(ts.URL+"/test",
-		req.Param{"param1": "abc", "param2": 123},
-		req.Header{
-			"Content-Type": "application/json",
-			"X-API-KEY":    "past-3",
-		},
-		req.BodyJSON(ExampleData2),
-	)
-	assert.NoError(t, err)
-	assert.True(t, publishCalled)
-}
-
 func TestGorillaGoMiddleware(t *testing.T) {
-	client := &Client{
-		config: &Config{
-			Debug:              true,
-			VerboseDebug:       true,
-			RedactHeaders:      []string{"X-Api-Key", "Accept-Encoding"},
-			RedactResponseBody: ExampleDataRedaction,
-		},
-	}
+	client := &apt.Client{}
+	client.SetConfig(&apt.Config{
+		Debug:              true,
+		VerboseDebug:       true,
+		RedactHeaders:      []string{"X-Api-Key", "Accept-Encoding"},
+		RedactResponseBody: apt.ExampleDataRedaction,
+	})
 	var publishCalled bool
-	client.PublishMessage = func(ctx context.Context, payload Payload) error {
+	client.PublishMessage = func(ctx context.Context, payload apt.Payload) error {
 		assert.Equal(t, "POST", payload.Method)
 		assert.Equal(t, "/{param1:[a-z]+}/test", payload.URLPath)
 		assert.Equal(t, map[string]string{"param1": "paramval"}, payload.PathParams)
@@ -119,10 +47,10 @@ func TestGorillaGoMiddleware(t *testing.T) {
 		assert.Equal(t, "/paramval/test?param1=abc&param2=123", payload.RawURL)
 		assert.Equal(t, http.StatusAccepted, payload.StatusCode)
 		assert.Greater(t, payload.Duration, 1000*time.Nanosecond)
-		assert.Equal(t, GoGorillaMux, payload.SdkType)
+		assert.Equal(t, apt.GoGorillaMux, payload.SdkType)
 
-		reqData, _ := json.Marshal(ExampleData2)
-		respData, _ := json.Marshal(ExampleDataRedacted)
+		reqData, _ := json.Marshal(apt.ExampleData2)
+		respData, _ := json.Marshal(apt.ExampleDataRedacted)
 
 		assert.Equal(t, reqData, payload.RequestBody)
 		assert.Equal(t, respData, payload.ResponseBody)
@@ -136,7 +64,7 @@ func TestGorillaGoMiddleware(t *testing.T) {
 		assert.NoError(t, err)
 		assert.NotEmpty(t, body)
 
-		jsonByte, err := json.Marshal(ExampleData)
+		jsonByte, err := json.Marshal(apt.ExampleData)
 		assert.NoError(t, err)
 
 		w.Header().Add("Content-Type", "application/json")
@@ -158,24 +86,23 @@ func TestGorillaGoMiddleware(t *testing.T) {
 			"Content-Type": "application/json",
 			"X-API-KEY":    "past-3",
 		},
-		req.BodyJSON(ExampleData2),
+		req.BodyJSON(apt.ExampleData2),
 	)
 	assert.NoError(t, err)
 	assert.True(t, publishCalled)
 }
 
 func TestOutgoingRequestGorilla(t *testing.T) {
-	client := &Client{
-		config: &Config{
-			Debug:              true,
-			VerboseDebug:       true,
-			RedactHeaders:      []string{"X-Api-Key", "Accept-Encoding"},
-			RedactResponseBody: ExampleDataRedaction,
-		},
-	}
+	client := &apt.Client{}
+	client.SetConfig(&apt.Config{
+		Debug:              true,
+		VerboseDebug:       true,
+		RedactHeaders:      []string{"X-Api-Key", "Accept-Encoding"},
+		RedactResponseBody: apt.ExampleDataRedaction,
+	})
 	var publishCalled bool
 	var parentId *string
-	client.PublishMessage = func(ctx context.Context, payload Payload) error {
+	client.PublishMessage = func(ctx context.Context, payload apt.Payload) error {
 		if payload.RawURL == "/from-gorilla" {
 			assert.NotNil(t, payload.ParentID)
 			parentId = payload.ParentID
@@ -209,7 +136,7 @@ func TestOutgoingRequestGorilla(t *testing.T) {
 			"Content-Type": "application/json",
 			"X-API-KEY":    "past-3",
 		},
-		req.BodyJSON(ExampleData2),
+		req.BodyJSON(apt.ExampleData2),
 	)
 	assert.NoError(t, err)
 	assert.True(t, publishCalled)
